backend/utility_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pytesseract==0.3.13
Pillow==11.1.0
pdfplumber==0.11.4
openpyxl==3.1.5
pandas==2.2.3
beautifulsoup4==4.12.3
feedparser==6.0.11
pyzbar==0.1.1
